feat: 評判の露出を実勢化(EXPOSURE_REFERENCE=2000)し、端数を累積する連続値に
100人で世論が動くのは低すぎるため、口コミ飽和点を実勢の来場規模へ。ただし露出を上げると
小集客の日は1日の前進が1点未満になり、整数丸めだと毎日消えてフリーズする。評判を連続値で
持ち、端数を翌日へ累積することで「世論は1日では動かないが、持続的な人気で動く」を表現する。

## 変更内容
- Reputation を連続値化: 内部 value(Float)を保持、score は丸めた整数(表示・需要用)。
  gain/lose/<=>/components を value ベースに、整数ガード→Numeric。
- ReputationPolicy: after_day を value から積み、drift の .round を廃止(端数累積)。
  EXPOSURE_REFERENCE 100→2,000(社会の口コミ飽和点。園の収容力とは無関係と明記)。
- 永続化: reputation 列 INTEGER→REAL、mapper を .value 保存(端数を往復で失わない)。
- テスト: CROWD=EXPOSURE_REFERENCE、露出比例テストは value 比較、Tier2 の満杯ケースは
  EXPOSURE_REFERENCE 露出に。17人/日は単日では動かない(51→50, よりリアル)。
  端数累積のガードテストを追加。

挙動: 露出17/日でも1年で50→59、200/日で数ヶ月、2000/日で1ヶ月で体験値へ収束。
評判の伝播速度が集客規模にスケールする。924 examples, 0 failures, 37 pending。

+++ lib/zoo/infrastructure/sqlite/database.rb
@@ -43,7 +43,7 @@ module Zoo
               revenue       INTEGER NOT NULL,
               visitor_count INTEGER NOT NULL,
               balance       INTEGER NOT NULL,
-              reputation    INTEGER NOT NULL,
+              reputation    REAL    NOT NULL,
               day           INTEGER NOT NULL DEFAULT 0
             );
             CREATE TABLE IF NOT EXISTS animals (
